extract slot component
