+++ src/routes/__layout.svelte
@@ -29,10 +29,8 @@
 	import {goto} from '$app/navigation';
 	import {PERSONA_QUERY_KEY, setUrlPersona} from '$lib/ui/url';
 	import Contextmenu from '$lib/ui/contextmenu/Contextmenu.svelte';
-	import SocketConnection from '$lib/ui/SocketConnection.svelte';
-	import {VITE_GIT_HASH} from '$lib/config';
+	import ContextmenuSlot from '$lib/app/ContextmenuSlot.svelte';
 	import {createContextmenuStore} from '$lib/ui/contextmenu/contextmenu';
-	import Avatar from '$lib/ui/Avatar.svelte';
 
 	let initialMobileValue = false; // TODO this hardcoded value causes mobile view to change on load -- detect for SSR via User-Agent?
 	const MOBILE_WIDTH = '50rem'; // treats anything less than 800px width as mobile
@@ -197,21 +195,6 @@
 		}
 	}
 
-	// TODO hacky, but if it sticks, upstream to Felt. maybe check things like `role="button"`?
-	export const isInteractive = (el: Element): boolean =>
-		el.tagName === 'A' || el.tagName === 'BUTTON' || !!el.closest('button,a');
-
-	const onClickContextmenuWrapper = (e: MouseEvent) => {
-		// this seems hacky, but improves the behavior to let us select content on the contextmenu,
-		// but automatically closes if e.g. a button is clicked, and the button can `stopPropagation`
-		// to keep the contextmenu open, because it'll stop it before this handler runs
-		if (isInteractive(e.target as any)) {
-			contextmenu.close();
-		} else {
-			e.stopPropagation();
-		}
-	};
-
 	$: layoutEntities = ['app', personaSelection ? 'persona:' + $personaSelection.name : '']
 		.filter(Boolean)
 		.join(',');
@@ -243,75 +226,7 @@
 	</main>
 	<Devmode {devmode} />
 	<Contextmenu {contextmenu}>
-		<!-- TODO refactor all of this -->
-		<!-- TODO maybe ignore Community if there's a Space? So it could combine into one view instead of 2 -->
-		<!-- TODO implement this for arbitrary items -- blocks? -->
-		<div class="contextmenu-wrapper" on:click={onClickContextmenuWrapper}>
-			{#each $contextmenu.entities as entity}
-				{#if $devmode}
-					<header class="panel-inset">{entity}</header>
-				{/if}
-				{#if entity === 'app'}
-					{#if $devmode}
-						<section>
-							<ul>
-								<li><a href="/docs">/docs</a></li>
-							</ul>
-						</section>
-						<section>
-							<SocketConnection />
-						</section>
-					{/if}
-					<section class="markup panel-inset">
-						<p>
-							<a href="https://github.com/feltcoop/felt-server" target="_blank" rel="noreferrer"
-								>felt-server</a
-							>
-							version 💚
-							<a
-								href="https://github.com/feltcoop/felt-server/commit/{VITE_GIT_HASH}"
-								target="_blank"
-							>
-								{VITE_GIT_HASH}
-							</a>
-						</p>
-					</section>
-				{:else if entity === 'luggage' || entity === 'selectedPersona'}
-					<section class="markup panel-outset">
-						<AccountForm guest={$session.guest} />
-					</section>
-					<!-- TODO refactor -->
-				{:else if entity.startsWith('persona:')}
-					<section class="markup panel-outset">
-						<Avatar name={entity.substring('persona:'.length)} />
-					</section>
-				{:else if entity.startsWith('community:')}
-					<section class="markup panel-outset">
-						<Avatar name={entity.substring('community:'.length)} type="Community" />
-					</section>
-				{:else if entity.startsWith('space:')}
-					<section class="markup panel-inset">
-						<h3>{entity.substring('space:'.length)}</h3>
-					</section>
-				{:else if entity.startsWith('file:')}
-					<section class="markup panel-inset">
-						<p>TODO use file_id: {entity.substring('file:'.length)}</p>
-					</section>
-				{:else if entity.startsWith('link:')}
-					<!-- TODO could do more if we had the original `target` element
-							(but it might go stale on $contextmenu?) -->
-					<!-- TODO if it's an external link, add target="_blank" -->
-					<a href={entity.substring('link:'.length)}>
-						<span class="icon">🔗</span>
-						{entity.substring('link:'.length)}
-					</a>
-				{:else}
-					<!-- <section class="markup">
-						<p>TODO default for entity: {entity}</p>
-					</section> -->
-				{/if}
-			{/each}
-		</div>
+		<ContextmenuSlot {contextmenu} {devmode} />
 	</Contextmenu>
 	<FeltWindowHost query={() => ({hue: randomHue($account?.name || GUEST_PERSONA_NAME)})} />
 </div>
@@ -331,46 +246,5 @@
 		align-items: center;
 		justify-content: center;
 		flex-direction: column;
-	}
-
-	.contextmenu-wrapper header {
-		text-align: center;
-		font-family: var(--font_family_mono);
-		font-size: var(--font_size_sm);
-		color: var(--text_color_light);
-		padding: var(--spacing_xs) 0;
-	}
-
-	.contextmenu-wrapper li a {
-		padding: var(--spacing_xs) var(--spacing_sm);
-		width: 100%;
-	}
-
-	.contextmenu-wrapper .markup {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-	}
-
-	.contextmenu-wrapper > section,
-	.contextmenu-wrapper > a {
-		border-bottom: var(--border);
-	}
-	.contextmenu-wrapper > section:last-child,
-	.contextmenu-wrapper > a:last-child {
-		border-bottom: none;
-	}
-
-	.contextmenu-wrapper > a {
-		display: flex;
-		align-items: center;
-		width: 100%;
-		word-break: break-word;
-	}
-
-	.icon {
-		display: flex;
-		font-size: var(--icon_size_sm);
-		padding: var(--spacing_sm);
 	}
 </style>
